# Relaxed contexts for HtmlT's Semigroup, Monoid, Functor, Applicative instances