The existing contexts were more restrictive than they needed to be.