@phadej 2018-02-07
@benjamin-hodgson could you run `bench-io` benchmark and paste its results with and without this patch?