@phadej 2018-02-07
Note to self: This is breaking change for pre-AMP GHCs