@benjamin-hodgson 2018-02-08
Before:
```
benchmarking renderText
time                 2.503 ms   (2.460 ms .. 2.550 ms)
                     0.998 R²   (0.997 R² .. 0.999 R²)
mean                 2.530 ms   (2.507 ms .. 2.552 ms)
std dev              79.24 μs   (65.61 μs .. 94.32 μs)
variance introduced by outliers: 16% (moderately inflated)

benchmarking renderTextT Identity
time                 2.565 ms   (2.531 ms .. 2.602 ms)
                     0.999 R²   (0.998 R² .. 0.999 R²)
mean                 2.578 ms   (2.554 ms .. 2.604 ms)
std dev              83.05 μs   (68.12 μs .. 113.4 μs)
variance introduced by outliers: 18% (moderately inflated)

benchmarking renderTextT Reader
time                 2.590 ms   (2.552 ms .. 2.623 ms)
                     0.998 R²   (0.997 R² .. 0.999 R²)
mean                 2.600 ms   (2.578 ms .. 2.623 ms)
std dev              77.73 μs   (67.23 μs .. 95.26 μs)
variance introduced by outliers: 16% (moderately inflated)

benchmarking renderTextT IO
time                 3.468 ms   (3.416 ms .. 3.509 ms)
                     0.998 R²   (0.997 R² .. 0.999 R²)
mean                 3.440 ms   (3.407 ms .. 3.480 ms)
std dev              115.3 μs   (93.74 μs .. 156.8 μs)
variance introduced by outliers: 16% (moderately inflated)
```

After:
```
benchmarking renderText
time                 2.868 ms   (2.831 ms .. 2.899 ms)
                     0.998 R²   (0.998 R² .. 0.999 R²)
mean                 2.818 ms   (2.788 ms .. 2.848 ms)
std dev              97.69 μs   (80.26 μs .. 130.1 μs)
variance introduced by outliers: 19% (moderately inflated)
             
benchmarking renderTextT Identity
time                 2.890 ms   (2.842 ms .. 2.933 ms)
                     0.998 R²   (0.997 R² .. 0.999 R²)
mean                 2.921 ms   (2.891 ms .. 2.953 ms)
std dev              104.7 μs   (82.64 μs .. 135.6 μs)
variance introduced by outliers: 19% (moderately inflated)
             
benchmarking renderTextT Reader
time                 2.894 ms   (2.851 ms .. 2.932 ms)
                     0.998 R²   (0.997 R² .. 0.999 R²)
mean                 2.937 ms   (2.907 ms .. 2.979 ms)
std dev              118.5 μs   (78.40 μs .. 181.3 μs)
variance introduced by outliers: 23% (moderately inflated)
             
benchmarking renderTextT IO
time                 3.175 ms   (3.146 ms .. 3.205 ms)
                     0.999 R²   (0.998 R² .. 0.999 R²)
mean                 3.132 ms   (3.106 ms .. 3.160 ms)
std dev              89.93 μs   (76.34 μs .. 107.7 μs)
variance introduced by outliers: 13% (moderately inflated)
```

So around 10%-15% slower for most of the benchmarks. Any idea what the cause of this might be? Do you think it's fixable?

Good point re AMP, I hadn't considered that. I guess this would need a revbump before release