@phadej 2018-02-08
@benjamin-hodgson could you do a PR with only relaxing `a ~ ()` => `Monoid a` (or `Semigroup`), also with benchmark results. It *shouldn't* affect performance and something I have thought previously

After that we can think what to do with `Applicative` stuff.