@phadej 2018-02-08
(Also note: you can see AMP effects in travis in 7.8 and older jobs)