@chrisdone 2018-02-08
You probably need the lazy ~ tildes in the places I put them. They improved performance at the time I wrote the code.