@benjamin-hodgson 2018-02-08
@phadej @chrisdone I put the lazy patterns back in (1afe5b9) and that fixed the performance problems. It now performs exactly the same as master.