@phadej 2018-02-08
@chrisdone one option is to drop support for pre-AMP GHC (7.8 and older). Otherwise we'd need somewhat ugly (`Applicative m, Monad m`) contexts (at least for old GHC.

And we do need, as travis shows.