@chrisdone 2018-02-08
I'm fine with dropping 7.8 and older, we could pop a min-bound on base and call it good.