@benjamin-hodgson 2018-02-08
Well, the immediate issue with Travis can be fixed by adjusting the `Monad` instance to not depend on the `Applicative` instance. Something like:

```haskell
pureInternal :: (a -> m a) -> a -> HtmlT m a
pureInternal pure = HtmlT $ pure $ (mempty, a)

instance Applicative m => Applicative (HtmlT m) where
    pure = pureInternal pure
    -- ...
instance Monad m => Monad (HtmlT m) where
    return = pureInternal return
    -- ...
```

That allows you to maintain compatibility with <7.10, though it's still a breaking change for users that have `Monad`s without `Applicative` instances.