@benjamin-hodgson 2018-02-08
Additional thought: the types of `makeElement`, `makeElementNoEnd` etc and the `Term`s which rely on them can probably also be relaxed to require `Applicative` rather than `Monad`.