@benjamin-hodgson 2018-02-08
I've been messing around with it a bit and I'm concluding that this was a bad idea after all.

You're right that the `Term` instances should use `Applicative` rather than `Monad`. However the `Monad` constraints in the `Term` instances come from `toHtml`. Since `toHtml` is an exported class method, changing that to `Applicative` would be a hard breaking change for anyone with their own `ToHtml` instances. (The instances in `Lucid.Base` only use `return` and not `>>=`, so they can be changed in principle, but users' instances can't.)

I'm going to close this PR. Thanks for the reviews, you were both very helpful!