@phadej 2018-02-08
@benjamin-hodgson are you sure. There is no need to change `ToHtml` class, `Term` has `Monad` context because of `With` (which could require only `Functor`), not `ToHtml`.