@benjamin-hodgson 2018-02-08
@phadej You were right, I was getting confused between `Term` and `TermRaw`.  I've implemented your suggestion, see 877aaf640e30758f50770d9357ca74f2b566ec6c