@benjamin-hodgson 2018-02-09
I got the build green, the code now works on pre-AMP GHCs (though it's still a breaking change). @chrisdone How do you feel about the `Monoid a` constraint on the `Monoid` instance? Would you like me to change it back to `~ ()`?