@chrisdone 2018-02-09
I'd prefer to leave it as `~ ()` unless you actually have a use-case for something else.