@benjamin-hodgson 2018-02-09
Done (c8f1b53). Can always be revisited in the future