@phadej 2018-02-09
@benjamin-hodgson thanks a lot!

@chrisdone I can do release in the next two days. I want to try this on our codebase, to see if it affects it at all.