@phadej 2018-03-09
This seems to work. I'll bump lower bound on base so we can simplify `Applicative m, Monad m` to bare `Monad m`.